chore: expose VmProviderConnection and VmProviderConnectionFactory  in extension-api Signed-off-by: Sonia Sandler <ssandler@redhat.com>
