+++ packages/main/src/plugin/provider-registry.spec.ts
@@ -34,6 +34,7 @@ import type {
   ProviderInstallation,
   ProviderLifecycle,
   ProviderUpdate,
+  VmProviderConnection,
 } from '@podman-desktop/api';
 import { assert, beforeEach, describe, expect, test, vi } from 'vitest';
 
@@ -49,7 +50,7 @@ import type { ApiSenderType } from './api.js';
 import type { AutostartEngine } from './autostart-engine.js';
 import type { ContainerProviderRegistry } from './container-registry.js';
 import { LifecycleContextImpl } from './lifecycle-context.js';
-import type { ProviderImpl, VmProviderConnection } from './provider-impl.js';
+import type { ProviderImpl } from './provider-impl.js';
 import { ProviderRegistry, type UpdateVmConnectionEvent } from './provider-registry.js';
 import type { Telemetry } from './telemetry/telemetry.js';
 import { Disposable } from './types/disposable.js';
